docs(BCardImg): Fill out component reference and migration guide
