+++ apps/docs/src/utils/string-utils.ts
@@ -1,0 +1,2 @@
+// Converts PascalCase or camelCase to kebab-case
+export const kebabCase = (str: string) => str.replace(/\B([A-Z])/g, '-$1').toLowerCase()
